Alter `createSpy` interface [BREAKING CHANGE]
Fortunately nobody uses this library yet, so it's not a big deal.

In bash, you can run:

     VAR=foo some_function

And the environment variable `VAR` is set only within `some_function`.

It turns out in other shells (like all of them), this syntax does not
scope the environment variable to just that function. To suggest to
people to use this syntax in non-Bash shells would be a horrible gotcha.

Hence I have updated the interface to expect options instead:

- `val=123 some_function somearg` → `some_function -r 123 somearg`
- `output=foo some_function somearg` → `some_function -o foo somearg`